+++ tests/terminaldetection/main.go
